#111 Move utils
